Refactor supported fields
